Remove support for deprecated startup properties
